Classes "GLmultiPhiGroup" "GLmultiPhiGroupSingle" "GLmultiFGroup" "GLmultiFGroupSingle" "GLCGMatrixGroupSingle"

Objects used to carry information mostly between invisible functions.

Objects from the Class

Objects can be created by calls of the form
new("GLmultiPhiGroup", MatriceCG, group = 'GLgroup', grindex = 'list')
new("GLmultiPhiGroupSingle", MatriceCG, group = 'GLgroup', grindex = 'list')
new("GLmultiFGroup", MatriceCG, group = 'GLgroup', grindex = 'list')
new("GLmultiFGroupSingle", MatriceCG, group = 'GLgroup', grindex = 'list')
new("GLCGMatrixGroupSingle", MatriceCG, group = 'GLgroup', grindex = 'list')

Slots

GLCGMatrixGroupSingle, GLmultiFGroupSingle, GLmultiPhiGroupSingle
.Data:

Object of class "matrix"

group:

Object of class "GLgroup"

grindex:

Object of class "list" containing the indices of the probands of 'group'

GLmultiPhiGroup :
.Data:

Object of class "GLmultiMatrix"

group:

Object of class "GLgroup"

grindex:

Object of class "list" containing the indices of the probands of 'group'

GLmultiFGroup :
.Data:

Object of class "GLmultiVector"

group:

Object of class "GLgroup"

grindex:

Object of class "list" containing the indices of the probands of 'group'

Extends

Class "matrix", from data part. Class "array", by class "matrix", distance 2. Class "structure", by class "matrix", distance 3. Class "vector", by class "matrix", distance 4, with explicit coerce.

Methods

[<-

signature(x = "GLCGMatrixGroupSingle", i = "ANY", j = "ANY", value = "ANY"): ...

[

signature(x = "GLCGMatrixGroupSingle", i = "ANY", j = "ANY", drop = "ANY"): ...

group

signature(x = "GLCGMatrixGroupSingle"): ...

Examples

showClass("GLCGMatrixGroupSingle")

#> Class "GLCGMatrixGroupSingle" [package "GENLIB"]
#> 
#> Slots:
#>                               
#> Name:    .Data   group grindex
#> Class:  matrix GLgroup    list
#> 
#> Extends: 
#> Class "matrix", from data part
#> Class "array", by class "matrix", distance 2
#> Class "mMatrix", by class "matrix", distance 2
#> Class "structure", by class "matrix", distance 3
#> Class "vector", by class "matrix", distance 4, with explicit coerce

showClass("GLmultiFGroup")

#> Class "GLmultiFGroup" [package "GENLIB"]
#> 
#> Slots:
#>                                       
#> Name:    .Data   group grindex   depth
#> Class:  matrix GLgroup    list integer
#> 
#> Extends: 
#> Class "GLmultiVector", directly
#> Class "matrix", by class "GLmultiVector", distance 2
#> Class "array", by class "GLmultiVector", distance 3
#> Class "mMatrix", by class "GLmultiVector", distance 3
#> Class "structure", by class "GLmultiVector", distance 4
#> Class "vector", by class "GLmultiVector", distance 5, with explicit coerce

showClass("GLmultiFGroupSingle")

#> Class "GLmultiFGroupSingle" [package "GENLIB"]
#> 
#> Slots:
#>                               
#> Name:    .Data   group grindex
#> Class:  matrix GLgroup    list
#> 
#> Extends: 
#> Class "matrix", from data part
#> Class "array", by class "matrix", distance 2
#> Class "mMatrix", by class "matrix", distance 2
#> Class "structure", by class "matrix", distance 3
#> Class "vector", by class "matrix", distance 4, with explicit coerce

showClass("GLmultiPhiGroupSingle")

#> Class "GLmultiPhiGroupSingle" [package "GENLIB"]
#> 
#> Slots:
#>                               
#> Name:    .Data   group grindex
#> Class:  matrix GLgroup    list
#> 
#> Extends: 
#> Class "matrix", from data part
#> Class "array", by class "matrix", distance 2
#> Class "mMatrix", by class "matrix", distance 2
#> Class "structure", by class "matrix", distance 3
#> Class "vector", by class "matrix", distance 4, with explicit coerce

showClass("GLmultiPhiGroup")

#> Class "GLmultiPhiGroup" [package "GENLIB"]
#> 
#> Slots:
#>                                       
#> Name:    .Data   group grindex   depth
#> Class:   array GLgroup    list integer
#> 
#> Extends: 
#> Class "GLmultiMatrix", directly
#> Class "array", by class "GLmultiMatrix", distance 2
#> Class "structure", by class "GLmultiMatrix", distance 3
#> Class "vector", by class "GLmultiMatrix", distance 4, with explicit coerce