Variance of completeness index

Returns the variance of the completeness index across probands.

gen.completenessVar( gen, pro = "0", genNo = -1, ...)

Arguments

gen
An object of class GLgen obtained with gen.genealogy, gen.lineages or gen.branching. Required.

pro
Vector of proband id numbers to be included. Default is 0, which will select all individuals without children.

genNo
Vector of generation numbers at which completeness should be calculated. Default is -1, which calculates completeness at each generation.

...
Option to pass additionnal arguments automaticaly between methods. Internal use only.

Value

A data frame with one column and each generation at which the variance is calculated as rows.

See also

References

Cazes P, Cazes MH. (1996) Comment mesurer la profondeur genealogique d'une ascendance? Population (French Ed) 51:117-140.

Examples

data(geneaJi) 
genJi<-gen.genealogy(geneaJi) 
gen.completeness(genJi, type="IND")

#>     Ind.1   Ind.2 Ind.29
#> 0 100.000 100.000 100.00
#> 1 100.000 100.000 100.00
#> 2 100.000 100.000 100.00
#> 3 100.000 100.000  75.00
#> 4  81.250  81.250  25.00
#> 5  50.000  50.000  12.50
#> 6  25.000  25.000   6.25
#> 7   3.125   3.125   0.00

gen.completenessVar(genJi)

#>   completeness.var
#> 0         0.000000
#> 1         0.000000
#> 2         0.000000
#> 3       208.333333
#> 4      1054.687500
#> 5       468.750000
#> 6       117.187500
#> 7         3.255208

# For the 4th generation
gen.completenessVar(genJi, genNo=4)

#>   completeness.var
#> 4         1054.688

# \donttest{
data(genea140) 
gen140<-gen.genealogy(genea140) 
gen.completeness(gen140)

#>            0            1            2            3            4            5 
#> 1.000000e+02 1.000000e+02 1.000000e+02 9.785714e+01 9.562500e+01 9.098214e+01 
#>            6            7            8            9           10           11 
#> 8.555804e+01 8.063616e+01 7.643415e+01 7.220145e+01 6.423828e+01 4.596261e+01 
#>           12           13           14           15           16           17 
#> 2.188825e+01 6.817976e+00 1.306763e+00 1.584298e-01 1.222883e-02 2.724784e-04 

gen.completenessVar(gen140)

#>    completeness.var
#> 0      0.000000e+00
#> 1      0.000000e+00
#> 2      0.000000e+00
#> 3      7.631038e+01
#> 4      1.661983e+02
#> 5      3.171490e+02
#> 6      5.774320e+02
#> 7      8.314375e+02
#> 8      9.573615e+02
#> 9      9.632942e+02
#> 10     8.953727e+02
#> 11     7.724232e+02
#> 12     4.428820e+02
#> 13     1.178367e+02
#> 14     9.342141e+00
#> 15     2.489817e-01
#> 16     2.251187e-03
#> 17     1.583512e-06

# }