Genetic contribution of ancestors

Returns the genetic contribution of ancestors to the gene pool of sepcifiec probands

gen.gc(gen, pro=0, ancestors=0, vctProb=c(0.5,0.5,0.5,0.5), typeCG="IND")

Arguments

gen
An object of class GLgen obtained with gen.genealogy, gen.lineages or gen.branching. Required.

pro
Vector of proband id numbers to be included. Default is 0, which will select all individuals without children.

ancestors
Vector of ancestors id numbers to be included. Default is 0, which will select all individuals without parents.

vctProb
Vector of transmission probabilities. The first two values indicate the probabilities of transmission of a father to his son and daughter, respectively, and the following two values are the same for the mother. Default is vctProb=c(0.5,0.5,0.5,0.5).

typeCG
IF typeCG="IND" (default), the genetic contribution from each ancestor is calculated for each proband. If typeCG="MEAN", the average (over all probands) genetic contributions of each ancestor is returned. If typeCG="PRODUCT", the product (over all probands) of genetic contributions is returned for each ancestor. If typeCG="TOTAL", the sum (over all probands) of genetic contributions is returned for each ancestor. If typeCG="CUMUL", ancestors are ranked in decreasing order of total contribution and cumulative contribution is returned.

Value

A matrix with rows corresponding to probands and columns corrsponding to ancestors when typeCG="IND". For the other typeCG values, rows are ancestors and there is one column containing mean, product, total or cumulative values.

See also

References

Roberts DF. (1968) Genetic effects of population size reduction. Nature, 220, 1084-1088.

O'Brien E, Jorde LB, Ronnlof B, Fellman JO, Eriksson AW. (1988) Founder effect and genetic disease in Sottunga, Finland. American Journal of Physical Anthropology, 77, 335-346.

Examples

data(geneaJi) 
genJi<-gen.genealogy(geneaJi) 
gc<-gen.gc(genJi, pro=c(1,29), ancestors=c(3,6,10,12,14,16,18,20,26,28))
gc

#>      3    6   10   12    14     16      18     20       26  28
#> 1  0.5 0.25 0.25 0.25 0.125 0.3125 0.21875 0.0625 0.109375 0.0
#> 29 0.0 0.00 0.00 0.00 0.250 0.1250 0.06250 0.2500 0.156250 0.5

gc_cum<-gen.gc(genJi, pro=c(1,29), ancestors=c(3,6,10,12,14,16,18,20,26,28), type="CUMUL")
gc_cum

#>       28        3       16       14       20       18       26       12 
#> 0.500000 1.000000 1.437500 1.812500 2.125000 2.406250 2.671875 2.921875 
#>       10        6 
#> 3.171875 3.421875 

data(genea140) 
gen140<-gen.genealogy(genea140) 
gc<-gen.gc(gen140, pro=c(454422,676521,677273,717537,717634,717709,868572), 
           ancestors=c(18305,18528,31114,18325))
gc

#>               18305        18528        31114       18325
#> 454422 0.0019531250 0.0020751953 0.0031738281 0.003173828
#> 676521 0.0000000000 0.0040283203 0.0009765625 0.005859375
#> 677273 0.0007629395 0.0008392334 0.0018920898 0.001098633
#> 717537 0.0002441406 0.0026855469 0.0018310547 0.010498047
#> 717634 0.0004882812 0.0018005371 0.0009765625 0.003906250
#> 717709 0.0001831055 0.0039978027 0.0006103516 0.015502930
#> 868572 0.0013427734 0.0022888184 0.0026855469 0.008300781

# Mother-daughter transmission only with probability=0.5
gc_MoLi<-gen.gc(gen140, pro=c(217891,302718,409282,802424,409682,443151),
    ancestors=c(18321,218231,296200,39066,18679,442607),vctProb=c(0,0,0,0.5))
gc_MoLi

#>        18321 218231 296200        39066        18679 442607
#> 217891     0      0      0 0.0000000000 0.0000000000 0.0000
#> 302718     0      0      0 0.0000000000 0.0000000000 0.0000
#> 409282     0      0      0 0.0000000000 0.0000000000 0.0000
#> 802424     0      0      0 0.0004882812 0.0000000000 0.0000
#> 409682     0      0      0 0.0000000000 0.0002441406 0.0000
#> 443151     0      0      0 0.0000000000 0.0000000000 0.0625