Variance of genealogical implex

Returns the variance of the implex index across probands.

gen.implexVar( gen, pro = "0", onlyNewAnc = FALSE, genNo = -1, ...)

Arguments

gen
An object of class GLgen obtained with gen.genealogy, gen.lineages or gen.branching. Required.

pro
Vector of proband id numbers to be included. Default is 0, which will select all individuals without children.

onlyNewAnc
If onlyNewAnc=FALSE (default), all ancestors will be considered. If onlyNewAnc=TRUE, only new ancestors will be counted (i.e., an ancestor is not counted again if it has already been counted in another generation).

genNo
Vector of generation numbers at which the implex should be calculated. Default is -1, which calculates the implex at each generation.

...
Additionnal arguments to be passed to methods.

Value

A data frame with one column and each generation at which the variance is calculated as rows

See also

References

Cazes P, Cazes MH. (1996) Comment mesurer la profondeur genealogique d'une ascendance? Population (French Ed) 51:117-140.

Examples

data(geneaJi) 
genJi<-gen.genealogy(geneaJi) 
gen.implex(genJi, type="IND")

#>      Ind.1    Ind.2  Ind.29
#> 0 100.0000 100.0000 100.000
#> 1 100.0000 100.0000 100.000
#> 2 100.0000 100.0000 100.000
#> 3  62.5000  62.5000  50.000
#> 4  31.2500  31.2500  12.500
#> 5  12.5000  12.5000   6.250
#> 6   6.2500   6.2500   3.125
#> 7   1.5625   1.5625   0.000

gen.implexVar(genJi)

#>    implex.var
#> 0   0.0000000
#> 1   0.0000000
#> 2   0.0000000
#> 3  52.0833333
#> 4 117.1875000
#> 5  13.0208333
#> 6   3.2552083
#> 7   0.8138021

# For the 5th generation
gen.implexVar(genJi, genNo=5)

#>   implex.var
#> 5   13.02083

# \donttest{
data(genea140) 
gen140<-gen.genealogy(genea140) 
gen.implex(gen140)

#>          implex
#> 0  1.000000e+02
#> 1  1.000000e+02
#> 2  1.000000e+02
#> 3  9.750000e+01
#> 4  9.428571e+01
#> 5  8.779018e+01
#> 6  7.939732e+01
#> 7  7.174665e+01
#> 8  6.384766e+01
#> 9  5.333845e+01
#> 10 4.018345e+01
#> 11 2.419120e+01
#> 12 9.309954e+00
#> 13 2.361101e+00
#> 14 3.948975e-01
#> 15 4.433768e-02
#> 16 3.749302e-03
#> 17 1.743862e-04

gen.implex(gen140, pro=c(454422, 676521, 677273, 717537, 717634, 717709, 868572), type="IND")

#>      Ind.454422   Ind.676521   Ind.677273   Ind.717537   Ind.717634
#> 0  100.00000000 1.000000e+02 1.000000e+02 100.00000000 100.00000000
#> 1  100.00000000 1.000000e+02 1.000000e+02 100.00000000 100.00000000
#> 2  100.00000000 1.000000e+02 1.000000e+02 100.00000000 100.00000000
#> 3   75.00000000 1.000000e+02 1.000000e+02 100.00000000  75.00000000
#> 4   75.00000000 1.000000e+02 1.000000e+02 100.00000000  50.00000000
#> 5   75.00000000 1.000000e+02 1.000000e+02 100.00000000  50.00000000
#> 6   75.00000000 1.000000e+02 9.375000e+01  96.87500000  46.87500000
#> 7   73.43750000 9.687500e+01 9.218750e+01  92.18750000  46.87500000
#> 8   67.57812500 8.828125e+01 8.984375e+01  84.76562500  44.53125000
#> 9   54.29687500 7.246094e+01 8.476562e+01  62.50000000  38.08593750
#> 10  41.30859375 5.380859e+01 8.007812e+01  48.53515625  26.46484375
#> 11  19.43359375 2.114258e+01 6.879883e+01  36.23046875  19.97070312
#> 12   4.29687500 3.735352e+00 5.100098e+01  13.42773438  10.25390625
#> 13   0.65917969 4.394531e-01 2.427979e+01   2.58789062   2.34375000
#> 14   0.06103516 2.441406e-02 6.024170e+00   0.37841797   0.32958984
#> 15   0.00000000 6.103516e-03 7.019043e-01   0.02441406   0.03051758
#> 16   0.00000000 0.000000e+00 4.577637e-02   0.00000000   0.00000000
#> 17   0.00000000 0.000000e+00 1.525879e-03   0.00000000   0.00000000
#>      Ind.717709   Ind.868572
#> 0  1.000000e+02 100.00000000
#> 1  1.000000e+02 100.00000000
#> 2  1.000000e+02 100.00000000
#> 3  1.000000e+02 100.00000000
#> 4  1.000000e+02 100.00000000
#> 5  9.375000e+01 100.00000000
#> 6  9.375000e+01 100.00000000
#> 7  9.375000e+01  97.65625000
#> 8  7.460938e+01  87.89062500
#> 9  4.843750e+01  64.64843750
#> 10 3.574219e+01  48.82812500
#> 11 2.451172e+01  34.52148438
#> 12 1.135254e+01  13.28125000
#> 13 3.076172e+00   3.19824219
#> 14 6.225586e-01   0.51269531
#> 15 6.713867e-02   0.04882812
#> 16 6.103516e-03   0.00000000
#> 17 0.000000e+00   0.00000000

gen.implexVar(gen140, pro=c(454422, 676521, 677273, 717537, 717634, 717709, 868572), type="IND")

#>      implex.var
#> 0  0.000000e+00
#> 1  0.000000e+00
#> 2  0.000000e+00
#> 3  1.488095e+02
#> 4  3.869048e+02
#> 5  3.701637e+02
#> 6  3.799293e+02
#> 7  3.446161e+02
#> 8  2.699207e+02
#> 9  2.398427e+02
#> 10 2.870487e+02
#> 11 3.091553e+02
#> 12 2.629978e+02
#> 13 7.180236e+01
#> 14 4.693623e+00
#> 15 6.512872e-02
#> 16 2.913709e-04
#> 17 3.326152e-07

# }