Expected Genealogical Depth

Returns the expected genealogical depth.

gen.meangendepth( gen, pro = "0", type = "MEAN", ...)

Arguments

gen
An object of class GLgen obtained with gen.genealogy, gen.lineages or gen.branching. Required.

pro
Vector of proband id numbers to be included. Default is 0, which will select all individuals without children.

type
If type="MEAN" (default), mean genealogical depth over all specified probands is calculated. If type="IND", mean genealogical depth is calculated for each specified proband.

...
Option to pass additionnal arguments automaticaly between methods. Internal use only.

Value

A data frame with only one numeric value when type is "MEAN". When type is "IND", the number of rows equals the number of probands specified.

See also

References

Cazes P, Cazes MH. (1996) Comment mesurer la profondeur genealogique d'une ascendance? Population (French Ed) 51:117-140.

Kouladjian K. (1986) Une mesure d'entropie genealogique. Chicoutimi, SOREP, Document III-C-43.

De Brakaeleer M, Bellis G. (1994) Genealogies et reconstitutions de familles en genetique humaine. Dossiers et Recherches, no 43, INED, Paris.

Examples

data(geneaJi) 
genJi<-gen.genealogy(geneaJi) 
gen.meangendepth(genJi)

#> [1] 4.125

gen.meangendepth(genJi, type="IND")

#>        Exp.Gen.Depth
#> Ind 1        4.59375
#> Ind 2        4.59375
#> Ind 29       3.18750

data(genea140) 
gen140<-gen.genealogy(genea140) 
gen.meangendepth(gen140)

#> [1] 9.396789

gen.meangendepth(gen140, pro=c(454422, 676521, 677273, 717537, 717634, 717709, 868572))

#> [1] 10.1408

gen.meangendepth(gen140, pro=c(454422, 676521, 677273, 717537, 717634, 717709, 868572), type="IND")

#>            Exp.Gen.Depth
#> Ind 454422      8.305298
#> Ind 676521     10.051086
#> Ind 677273     11.661545
#> Ind 717537     11.433716
#> Ind 717634      6.917358
#> Ind 717709     11.472717
#> Ind 868572     11.143860