Ancestor occurences

Returns the number of times that the specified ancestors are present in the genalogies of the specified probands.

gen.occ( gen, pro = "0", ancestors = "0", typeOcc = "IND", ...)

Arguments

gen
An object of class GLgen obtained with gen.genealogy, gen.lineages or gen.branching. Required.

pro
Vector of proband id numbers to be included. Default is 0, which will select all individuals without children.

ancestors
Vector of ancestors id numbers to be included. Default is 0, which will select all individuals without parents.

typeOcc
If typeOcc="IND" (default), the number of occurences for each proband will be returned. If typeOcc="TOTAL", the sum of all occurences over all probands will be returned.

...
Option to pass additionnal arguments automaticaly between methods. Internal use only.

Value

A matrix with number of lines equal to the number of ancestors included and the number of columns equal to the number of probands included if typeOcc="BRUT" or only one column if typeOcc="TOTAL"

See also

Examples

data(geneaJi) 
genJi<-gen.genealogy(geneaJi) 
# Number of occurences of ancestors in the specified proband's genealogy
gen.occ(genJi, pro=c(1,29), ancestors=c(3,6,10,12,14,16,18,20,26,28))

#>    1 29
#> 3  1  0
#> 6  1  0
#> 10 2  0
#> 12 2  0
#> 14 2  2
#> 16 6  2
#> 18 8  2
#> 20 1  2
#> 26 8  3
#> 28 0  1

data(genea140) 
gen140<-gen.genealogy(genea140) 
gen.occ(gen140, pro=c(454422,676521,677273,717537,717634,717709,868572), 
        ancestors=c(18305,18528,31114,18325))

#>       454422 676521 677273 717537 717634 717709 868572
#> 18305      3      0      8      1      1      2      4
#> 18528     13     17     21     41     25     52     30
#> 31114      4      1     14      7      4      2      8
#> 18325      6      7     12     41     17     51     29